Add support for logging in as a different user
This commit adds a new option for selecting what user to log in to
the container as. The login mechanism works the same as before, but
allows it to be setup for a user other than root.

If the user does not exist or the account is non-login, nixos-shell
will print the error message "This account is currently not available"
and exit with status code 1.